Week_8/Day5/deploy/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
llama-cpp-python==0.3.2
streamlit==1.38.0
requests==2.32.3